radio_registers: Changed it to temporarily use old settings (from revision test2/radio_mac.c in revision 2409 of Pololu's private repository), and the packet error rate I was getting in the wireless_serial app got WAY better.  Effective data rate went up by a factor of ~40 and is now about 12000 bytes per second!  Also, in radio_mac and radio_link changed the units of the timeout to be .998 ms, so that a timeout of 1 is acceptable and we can avoid doing 16-bit arithmetic.

+++ libraries/include/radio_mac.h
@@ -37,7 +37,7 @@ extern volatile uint8 DATA radioMacState;
 
 // State-changing functions.  These should not be called if radioMacState == RADIO_MAC_STATE_TX.
 void radioMacTx(uint8 XDATA * packet);
-void radioMacRx(uint8 XDATA * packet, uint16 timeout);
+void radioMacRx(uint8 XDATA * packet, uint8 timeout);
 
 // This function is called by the Mac in an ISR when an important event
 // happens (and puts the MAC in the idle state).  This function should be
